requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
confusable-homoglyphs==3.2.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filer==2.0.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-registration==3.1.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.7.1
idna==2.10
mysql-connector-python==8.0.23
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.2.0
protobuf==3.15.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.2
urllib3==1.26.4
mojimoji~=0.0.12